DAT77: Rename NetHandler to ServerNetHandler
