+++ client/agent/user/scanuser.vcxproj
@@ -173,8 +173,9 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="NetHandler.c" />
+    <ClCompile Include="ServerNetHandler.c" />
     <ClCompile Include="scanUser.c" />
+    <ClCompile Include="UINetHandler.c" />
     <ClCompile Include="Utils.c" />
     <ResourceCompile Include="scanUser.rc" />
   </ItemGroup>
@@ -188,7 +189,8 @@
     <None Exclude="@(None)" Include="*.def;*.bat;*.hpj;*.asmx" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Exclude="@(ClInclude)" Include="*.h;*.hpp;*.hxx;*.hm;*.inl;*.xsd" />
+	<ClInclude Exclude="@(ClInclude)" Include="ServerNetHandler.h" />
+	<ClInclude Include="UINetHandler.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
 </Project>
